requirements.txt
django==2.2
django-ckeditor==5.9.0
djangorestframework==3.10.3
Pillow==6.0.0
django-crispy-forms==1.7.2
social-auth-app-django==3.1.0